+++ TeadsSampleApp/Controllers/Native/AppLovin/TableView/AppLovinNativeAdView.swift
@@ -17,12 +17,3 @@ import MoPubSDK
 class AppLovinNativeAdView: MANativeAdView {
     
 }
-
-extension UIView {
-    static func loadNib(nibName name: String? = nil) -> Self? {
-        let nibName = name ?? String(describing: Self.self)
-        let bundle = Bundle(for: Self.self)
-        let nib = UINib(nibName: nibName, bundle: bundle)
-        return nib.instantiate(withOwner: self, options: nil).first as? Self
-    }
-}
